+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/Management/Action/NtiUnderConsideration/Edit.cshtml.cs
@@ -126,7 +126,7 @@ namespace ConcernsCaseWork.Pages.Case.Management.Action.NtiUnderConsideration
 
 			return reasonValues.Select(r => new RadioItem
 			{
-				Id = Convert.ToString(r),
+				Id = Convert.ToString((int)r),
 				Text = EnumHelper.GetEnumDescription(r),
 				IsChecked = ntiModel?.NtiReasonsForConsidering?.Any(ntiR => ntiR.Id == (int)r) == true,
 			});
